backend/foodgram/requirements.txt
Django==2.2.19
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.20.0
django-extensions==3.1.5
ipython==7.34.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
django-filter==21.1
pillow==9.1.1
sorl-thumbnail==12.8.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
reportlab==3.6.9
